Refactor: split down code
