+++ src/styles.py
@@ -1,0 +1,3 @@
+from rich.style import Style
+
+HIGHLIGHT = Style(bgcolor="bright_black")
